Fix crash when exiting a partially-initialized window.
[ ] May break build
[ ] Breaks existing application
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation

+++ examples/eqPly/window.cpp
@@ -52,7 +52,9 @@ bool Window::configInit( const uint32_t initID )
 
 bool Window::configExit()
 {
-    _state->deleteAll();
+    if( _state )
+        _state->deleteAll();
+
     delete _state;
     _state = 0;
